* commit 'trunk': (83 commits)
  Stop direct loading of files in wp-admin that should only be included, for trunk
  More cap checks, for trunk
  Add some CYA cap checks.
  Add some CYA cap checks.
  Be consistent about slashing _REQUEST superglobal. props dd32. fixes #10360
  Force IPv4 (127.0.0.1) when connecting to localhost, as many dev environments are not set up to serve on IPv6 (::1). props dd32. fixes #10045
  add some filters so that HTTP requests can be filtered. props johnbillion. fixes #10413
  Fix some default widget PHP notices. props mrmist. fixes #10225
  Return an empty array instead of false for get_children() when no children found. props Denis-de-Bernardy. fixes #10053
  Add missing quote. Props brianwhite. fixes #10474
  Fix invalid border-radius css, props Simek, fixes #10500 for trunk
  Don't show edit link for comments when the user cannot edit them, props filosofo, fixes #10520 for trunk
  Trash status update, see #4529
  Trash status updates for posts, pages, comments and attachments, props caesarsgrunt, see #4529
  Allow _wp_get_comment_list to handle custom comment types, props aaroncampbell, fixes #10463
  Add attachment ID to the wp_generate_attachment_metadata filter, props taco1991, fixes #10494
  Use exif rotation data when creating sub-sizes of uploaded jpeg images, see #7042
  Canonical redirects for post name queries.  Props DD32. see #6516
  Fix wp_tag_cloud('order=RAND'), don't pre-sort if filter 'tag_cloud_sort' is used to sort the tags, props Joern_W, fixes #10393
  Improve QE data for Draft, Pending and Scheduled posts screens, see #10478
  ...

+++ wp-includes/comment.php
@@ -208,6 +208,8 @@ function get_comments( $args = '' ) {
 		$approved = "comment_approved = '1'";
 	elseif ( 'spam' == $status )
 		$approved = "comment_approved = 'spam'";
+	elseif ( 'trash' == $status )
+		$approved = "comment_approved = 'trash'";
 	else
 		$approved = "( comment_approved = '0' OR comment_approved = '1' )";
 
@@ -692,14 +694,15 @@ function wp_count_comments( $post_id = 0 ) {
 	if ( false !== $count )
 		return $count;
 
-	$where = '';
+	$where = 'WHERE ';
 	if( $post_id > 0 )
-		$where = $wpdb->prepare( "WHERE comment_post_ID = %d", $post_id );
+		$where .= $wpdb->prepare( "c.comment_post_ID = %d AND ", $post_id );
+	$where .= "p.post_status <> 'trash'";
 
-	$count = $wpdb->get_results( "SELECT comment_approved, COUNT( * ) AS num_comments FROM {$wpdb->comments} {$where} GROUP BY comment_approved", ARRAY_A );
+	$count = $wpdb->get_results( "SELECT comment_approved, COUNT( * ) AS num_comments FROM {$wpdb->comments} c LEFT JOIN {$wpdb->posts} p ON c.comment_post_ID = p.ID {$where} GROUP BY comment_approved", ARRAY_A );
 
 	$total = 0;
-	$approved = array('0' => 'moderated', '1' => 'approved', 'spam' => 'spam');
+	$approved = array('0' => 'moderated', '1' => 'approved', 'spam' => 'spam', 'trash' => 'trash');
 	$known_types = array_keys( $approved );
 	foreach( (array) $count as $row_num => $row ) {
 		$total += $row['num_comments'];
@@ -736,9 +739,19 @@ function wp_count_comments( $post_id = 0 ) {
  */
 function wp_delete_comment($comment_id) {
 	global $wpdb;
-	do_action('delete_comment', $comment_id);
+	if (!$comment = get_comment($comment_id))
+		return false;
 
-	$comment = get_comment($comment_id);
+	if (wp_get_comment_status($comment_id) != 'trash' && wp_get_comment_status($comment_id) != 'spam' && EMPTY_TRASH_DAYS > 0)
+		return wp_trash_comment($comment_id);
+	
+	do_action('delete_comment', $comment_id);
+	
+	$trash_meta = get_option('wp_trash_meta');
+	if (is_array($trash_meta) && isset($trash_meta['comments'][$comment_id])) {
+		unset($trash_meta['comments'][$comment_id]);
+		update_option('wp_trash_meta', $trash_meta);
+	}
 
 	if ( ! $wpdb->query( $wpdb->prepare("DELETE FROM $wpdb->comments WHERE comment_ID = %d LIMIT 1", $comment_id) ) )
 		return false;
@@ -762,12 +775,72 @@ function wp_delete_comment($comment_id) {
 }
 
 /**
+ * Moves a comment to the Trash
+ *
+ * @since 2.9.0
+ * @uses do_action() on 'trash_comment' before trashing
+ * @uses do_action() on 'trashed_comment' after trashing
+ *
+ * @param int $comment_id Comment ID.
+ * @return mixed False on failure
+ */
+function wp_trash_comment($comment_id = 0) {
+	if (EMPTY_TRASH_DAYS == 0)
+		return wp_delete_comment($comment_id);
+
+	if (!$comment = get_comment($comment_id))
+		return false;
+
+	do_action('trash_comment', $comment_id);
+
+	$trash_meta = get_option('wp_trash_meta', array());
+	$trash_meta['comments'][$comment_id]['status'] = $comment->comment_approved;
+	$trash_meta['comments'][$comment_id]['time'] = time();
+	update_option('wp_trash_meta', $trash_meta);
+
+	wp_set_comment_status($comment_id, 'trash');
+
+	do_action('trashed_comment', $comment_id);
+
+	return true;
+}
+
+/**
+ * Removes a comment from the Trash
+ *
+ * @since 2.9.0
+ * @uses do_action() on 'untrash_comment' before undeletion
+ * @uses do_action() on 'untrashed_comment' after undeletion
+ *
+ * @param int $comment_id Comment ID.
+ * @return mixed False on failure
+ */
+function wp_untrash_comment($comment_id = 0) {
+	do_action('untrash_comment', $comment_id);
+
+	$comment = array('comment_ID'=>$comment_id, 'comment_approved'=>'0');
+
+	$trash_meta = get_option('wp_trash_meta');
+	if (is_array($trash_meta) && isset($trash_meta['comments'][$comment_id])) {
+		$comment['comment_approved'] = $trash_meta['comments'][$comment_id]['status'];
+		unset($trash_meta['comments'][$comment_id]);
+		update_option('wp_trash_meta', $trash_meta);
+	}
+
+	wp_update_comment($comment);
+
+	do_action('untrashed_comment', $comment_id);
+
+	return true;
+}
+
+/**
  * The status of a comment by ID.
  *
  * @since 1.0.0
  *
  * @param int $comment_id Comment ID
- * @return string|bool Status might be 'deleted', 'approved', 'unapproved', 'spam'. False on failure.
+ * @return string|bool Status might be 'trash', 'approved', 'unapproved', 'spam'. False on failure.
  */
 function wp_get_comment_status($comment_id) {
 	$comment = get_comment($comment_id);
@@ -777,13 +850,15 @@ function wp_get_comment_status($comment_id) {
 	$approved = $comment->comment_approved;
 
 	if ( $approved == NULL )
-		return 'deleted';
+		return false;
 	elseif ( $approved == '1' )
 		return 'approved';
 	elseif ( $approved == '0' )
 		return 'unapproved';
 	elseif ( $approved == 'spam' )
 		return 'spam';
+	elseif ( $approved == 'trash' )
+		return 'trash';
 	else
 		return false;
 }
@@ -983,7 +1058,7 @@ function wp_new_comment( $commentdata ) {
 	$commentdata['comment_parent'] = ( 'approved' == $parent_status || 'unapproved' == $parent_status ) ? $commentdata['comment_parent'] : 0;
 
 	$commentdata['comment_author_IP'] = preg_replace( '/[^0-9a-fA-F:., ]/', '',$_SERVER['REMOTE_ADDR'] );
-	$commentdata['comment_agent']     = $_SERVER['HTTP_USER_AGENT'];
+	$commentdata['comment_agent']     = substr($_SERVER['HTTP_USER_AGENT'], 0, 254);
 
 	$commentdata['comment_date']     = current_time('mysql');
 	$commentdata['comment_date_gmt'] = current_time('mysql', 1);
@@ -1028,7 +1103,7 @@ function wp_new_comment( $commentdata ) {
  */
 function wp_set_comment_status($comment_id, $comment_status, $wp_error = false) {
 	global $wpdb;
-
+		
 	$status = '0';
 	switch ( $comment_status ) {
 		case 'hold':
@@ -1044,8 +1119,8 @@ function wp_set_comment_status($comment_id, $comment_status, $wp_error = false) 
 		case 'spam':
 			$status = 'spam';
 			break;
-		case 'delete':
-			return wp_delete_comment($comment_id);
+		case 'trash':
+			$status = 'trash';
 			break;
 		default:
 			return false;
@@ -1324,6 +1399,9 @@ function do_all_pings() {
 	if ( is_array($trackbacks) )
 		foreach ( $trackbacks as $trackback )
 			do_trackbacks($trackback);
+
+	//Do Update Services/Generic Pings
+	generic_ping();
 }
 
 /**
